added overloaded method for querying the patient database.
fireTableDataChanged was overloaded to accept a String argument.  This
argument is the query used when querying the database.  Omitting the
argument results in a call to the zero-argument method, executing the
query "SELECT * FROM patient"

This will add a lot of flexibility to the application, and has been
used to implement searching and the "in hospital" button.